==> 01-Creating-Database.Rmd <==
# Creating a database

Below I display how to create a database for my final project in reproducible data science for 2023. 
Provided is an example databse that we will be creating.

```{r, echo=FALSE}
knitr::include_graphics("ThesisDatabase.png")
```

Load all packages needed to create the database

```{r, eval=FALSE}
library(DBI)
```

## Establish the connection with database
library(DBI)
Here we will establish the connection for the database to the directory using `dbconnect`

```{r, eval=FALSE}
Thesis_database <- dbConnect(RSQLite::SQLite(), 
                        "C:/Users/jaran/Desktop/R/Spr22Classwork/Reproducible Science/Final_Project/SQL_database/Thesis_Database.db")

class(Thesis_database)
```

## Creating Tables

Below I will create a table using the `dbexecute` function. Within this function I will then use SQL language to create tables for my database.

```{r, eval=FALSE}
dbExecute(Thesis_database, "CREATE TABLE Lion (
                                                lion_id varchar(6) NOT NULL,
                                                sex char(1) CHECK (sex IN ('M', 'F')),
                                                age char(2),
                                                capture_date text,
                                                capture_site varchar(25),
                                                PRIMARY KEY (lion_id));")

dbExecute(Thesis_database, "CREATE TABLE Capture (
                                                lion_id varchar(6) NOT NULL,
                                                capture_id char(4),
                                                capture_site varchar(25),
                                                capture_date text,
                                                utm_x  float,
                                                utm_y  float,
                                                PRIMARY KEY (capture_id)
                                                FOREIGN KEY (capture_site) REFERENCES Lion(capture_site)
                                                FOREIGN KEY (lion_id) REFERENCES Lion(lion_id)
                                                FOREIGN KEY (capture_date) REFERENCES Lion(capture_date)
                                                );")

dbExecute(Thesis_database, "CREATE TABLE GPS_Collar (
                                                     lion_id varchar(6) NOT NULL,
                                                    lionid_collarid varchar(25)),
                                                     collar_id varchar(10),
                                                     make varchar(25),
                                                     model varchar(25),
                                                     deployment_start text,
                                                     deploymeny_end text,
                                                     PRIMARY KEY (lionid_collarid)
                                                     FOREIGN KEY (lion_id) REFERENCES Lion(lion_id)
                                                     );")

dbExecute(Thesis_database, "CREATE TABLE GPS_Locaton (
                                                     locationid varchar(20) NOT NULL,
                                                    lionid_collarid varchar(25),
                                                     collar_id varchar(10),
                                                     localtime text,
                                                     utme_x float,
                                                     utmn_y float,
                                                     PRIMARY KEY (locationid)
                                                     FOREIGN KEY (lionid_collarid) REFERENCES

```
